layout/generic/test/test_plugin_clipping2.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="/tests/SimpleTest/test.css" type="text/css"?>
<html xmlns="http://www.w3.org/1999/xhtml" title="Test Plugin Clipping: Dynamic Tests">
<head>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>

<script class="testbody" type="application/javascript">
<![CDATA[
if (navigator.platform.indexOf("Mac") == 0) {
  todo(false, "Mac plugins do not support complex clipping");
} else {
  window.open("plugin_clipping_helper2.xhtml", "", "width=620,height=320");
  SimpleTest.waitForExplicitFinish();
}
]]>
</script>

</body>
</html>
